Rename view to ui
