+++ app/src/main/kotlin/com/ayatk/biblio/ui/activity/SearchActivity.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017. Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.view.activity
+package com.ayatk.biblio.ui.activity
 
 import android.content.Context
 import android.content.Intent
@@ -24,8 +24,8 @@ import com.ayatk.biblio.R
 import com.ayatk.biblio.R.layout
 import com.ayatk.biblio.databinding.ActivitySearchBinding
 import com.ayatk.biblio.databinding.ViewSearchResultItemBinding
-import com.ayatk.biblio.view.customview.BindingHolder
-import com.ayatk.biblio.view.customview.ObservableListRecyclerAdapter
+import com.ayatk.biblio.ui.customview.BindingHolder
+import com.ayatk.biblio.ui.customview.ObservableListRecyclerAdapter
 import com.ayatk.biblio.viewmodel.SearchResultItemViewModel
 import com.ayatk.biblio.viewmodel.SearchViewModel
 import io.reactivex.android.schedulers.AndroidSchedulers
