+++ app/src/main/kotlin/com/ayatk/biblio/ui/customview/ObservableListRecyclerAdapter.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.view.customview
+package com.ayatk.biblio.ui.customview
 
 import android.content.Context
 import android.databinding.ObservableList
