+++ app/src/main/kotlin/com/ayatk/biblio/ui/fragment/BaseFragment.kt
@@ -2,12 +2,12 @@
  * Copyright (c) 2016-2017 Aya Tokikaze. All Rights Reserved.
  */
 
-package com.ayatk.biblio.view.fragment
+package com.ayatk.biblio.ui.fragment
 
 import android.support.v4.app.Fragment
 import com.ayatk.biblio.di.FragmentComponent
 import com.ayatk.biblio.di.FragmentModule
-import com.ayatk.biblio.view.activity.BaseActivity
+import com.ayatk.biblio.ui.activity.BaseActivity
 
 open class BaseFragment : Fragment() {
