+++ app/src/main/res/layout/activity_novel_detail.xml
@@ -7,7 +7,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.activity.NovelDetailActivity">
+    tools:context=".ui.activity.NovelDetailActivity">
 
   <data>
