+++ app/src/main/res/layout/activity_search.xml
@@ -7,7 +7,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".view.activity.SearchActivity">
+    tools:context=".ui.activity.SearchActivity">
 
   <android.support.v4.widget.DrawerLayout
       android:id="@+id/drawer_layout"
