+++ build.gradle
@@ -15,7 +15,7 @@ buildscript {
   }
 
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.0.0'
+    classpath 'com.android.tools.build:gradle:3.0.1'
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:${kotlin_version}"
     classpath 'com.google.gms:google-services:3.1.0'
     classpath 'com.cookpad.android.licensetools:license-tools-plugin:0.23.0'
